+++ src/DotNetWorker.ApplicationInsights/release_notes.md
@@ -1,7 +1,5 @@
 ## What's Changed
 
-### Microsoft.Azure.Functions.Worker.ApplicationInsights <version>
+### Microsoft.Azure.Functions.Worker.ApplicationInsights 2.51.0
 
-- Updating `Azure.Identity` from 1.12.0 to 1.17.0
-- Updating `Microsoft.ApplicationInsights.PerfCounterCollector` from 2.22.0 to 2.23.0
 - Improved idempotency of service registration calls (#3273)